LayoutTests/http/tests/site-isolation/inspector/runtime/evaluate-in-cross-origin-iframe.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<script src="../../../inspector/resources/inspector-test.js"></script>
<script src="../debugger/resources/site-isolation-debugger-test-utilities.js"></script>
<script>
    var passphrase = "main-page-value";

    function test() {
        let suite = InspectorTest.createAsyncSuite("SiteIsolation.Runtime.CrossOriginEvaluate");

        let crossOriginTarget;

        suite.addTestCase({
            name: "SiteIsolation.Runtime.CrossOrigin.FrameTargetCreated",
            description: "The static cross-origin iframe should have a frame target with an execution context.",
            async test() {
                let frameTargets = WI.targets.filter((t) => t.type === WI.TargetType.Frame);
                let urlTargetMap = await fetchDocumentURLsForTargets(frameTargets);
                let mainFrameTarget = urlTargetMap.get(WI.networkManager.mainFrame.url);
                crossOriginTarget = frameTargets.find((t) => t !== mainFrameTarget);

                InspectorTest.assert(crossOriginTarget instanceof WI.FrameTarget);
                InspectorTest.expectNotNull(crossOriginTarget.RuntimeAgent, "Cross-origin frame target should have a RuntimeAgent.");
                InspectorTest.expectNotNull(crossOriginTarget.executionContext, "Cross-origin frame target should have an ExecutionContext.");
            }
        });

        suite.addTestCase({
            name: "SiteIsolation.Runtime.CrossOrigin.EvaluatePassphrase",
            description: "Evaluating in the cross-origin frame should return values from that frame's context.",
            async test() {
                let passphraseValue = await waitForExpressionInTarget(crossOriginTarget, "window.passphrase");
                InspectorTest.expectEqual(passphraseValue, "cross-origin-secret", "Cross-origin frame should have passphrase 'cross-origin-secret'.");

                let pageResponse = await WI.pageTarget.RuntimeAgent.evaluate.invoke({ expression: "window.passphrase", objectGroup: "test", returnByValue: true });
                InspectorTest.expectEqual(pageResponse.result.value, "main-page-value", "Main page should have passphrase 'main-page-value'.");

                InspectorTest.expectNotEqual(passphraseValue, pageResponse.result.value, "Cross-origin frame and main page should have different passphrase values.");
            }
        });

        suite.addTestCase({
            name: "SiteIsolation.Runtime.CrossOrigin.EvaluateHostname",
            description: "Evaluating document.location.hostname should return the cross-origin hostname.",
            async test() {
                let frameHostname = await waitForExpressionInTarget(crossOriginTarget, "document.location.hostname");
                let pageResponse = await WI.pageTarget.RuntimeAgent.evaluate.invoke({ expression: "document.location.hostname", objectGroup: "test", returnByValue: true });

                InspectorTest.expectNotEqual(frameHostname, pageResponse.result.value, "Cross-origin frame hostname should differ from main page hostname.");
                InspectorTest.log(`Main page hostname: ${pageResponse.result.value}`);
                InspectorTest.log(`Cross-origin frame hostname: ${frameHostname}`);
            }
        });

        suite.addTestCase({
            name: "SiteIsolation.Runtime.CrossOrigin.CallFunctionOn",
            description: "callFunctionOn should work with objects created in the cross-origin frame's context.",
            async test() {
                let createResponse = await crossOriginTarget.RuntimeAgent.evaluate.invoke({ expression: "({key: 'from-cross-origin'})", objectGroup: "test" });
                InspectorTest.expectThat(!createResponse.wasThrown, "Creating object in cross-origin frame should not throw.");
                InspectorTest.expectNotNull(createResponse.result.objectId, "Created object should have an objectId.");

                let callResponse = await crossOriginTarget.RuntimeAgent.callFunctionOn.invoke({
                    objectId: createResponse.result.objectId,
                    functionDeclaration: "function() { return this.key; }",
                    returnByValue: true,
                });
                InspectorTest.expectEqual(callResponse.result.value, "from-cross-origin", "callFunctionOn should return the object's key value.");
                InspectorTest.expectThat(!callResponse.wasThrown, "callFunctionOn should not throw.");
            }
        });

        suite.runTestCasesAndFinish();
    }
</script>

<body onload="runTest()">
    <p>Test that Runtime.evaluate and callFunctionOn work correctly in cross-origin iframes under site isolation.</p>
    <iframe src="http://localhost:8000/site-isolation/inspector/runtime/resources/frame-with-passphrase.html"></iframe>
</body>
